double->float

+++ Lattice.cpp
@@ -131,28 +131,28 @@ void Lattice_3D::create_lattice() {
     std::cout << "Finish lattice creation " << std::endl;
 }
 
-double Lattice_2D::radius(const coord_t& start, const coord_t& end) {
+float Lattice_2D::radius(const coord_t& start, const coord_t& end) {
     long start_x = start % lattice_side;
     long start_y = start / lattice_side;
     long end_x = end % lattice_side;
     long end_y = end / lattice_side;
 
     //torus distance;
-    double xdiff = abs(end_x - start_x);
+    float xdiff = abs(end_x - start_x);
     if (xdiff > (lattice_side/2))
         xdiff = lattice_side - xdiff;
 
-    double ydiff = abs(end_y - start_y);
+    float ydiff = abs(end_y - start_y);
     if (ydiff > (lattice_side / 2))
         ydiff = lattice_side - ydiff;
 
-    double r = xdiff *xdiff  + ydiff*ydiff ;
+    float r = xdiff *xdiff  + ydiff*ydiff ;
 
     return r;
 }
 
 KOKKOS_INLINE_FUNCTION
-double Lattice_3D::radius(const coord_t& start, const coord_t& end) {
+float Lattice_3D::radius(const coord_t& start, const coord_t& end) {
     long start_x = start % lattice_side;
     long start_y = (start % (lattice_side * lattice_side)) /lattice_side;
     long start_z = start / (lattice_side * lattice_side);
@@ -161,19 +161,19 @@ double Lattice_3D::radius(const coord_t& start, const coord_t& end) {
     long end_z = end / (lattice_side * lattice_side);
 
     //torus distance;
-    double xdiff = abs(end_x - start_x);
+    float xdiff = abs(end_x - start_x);
     if (xdiff > (lattice_side/2))
         xdiff = lattice_side - xdiff;
 
-    double ydiff = abs(end_y - start_y);
+    float ydiff = abs(end_y - start_y);
     if (ydiff > (lattice_side / 2))
         ydiff = lattice_side - ydiff;
 
-    double zdiff = abs(end_z - start_z);
+    float zdiff = abs(end_z - start_z);
     if (zdiff > (lattice_side / 2))
         zdiff = lattice_side - zdiff;
 
-    double r = xdiff *xdiff  + ydiff*ydiff + zdiff*zdiff;
+    float r = xdiff *xdiff  + ydiff*ydiff + zdiff*zdiff;
 
     return r;
 }
